refactor(guestbook): guestbook 이름 및 이름 리팩토링
- guestbook UI 컴포넌트 이동 및 이름 변경으로 구조 정리
- ThemeProvider를 ThemeContext로 이름 변경하고 관련 import 갱신
- 윈도우 설정 파일 이름 변경 및 참조 경로 정리
- window.types.ts에 Position 타입 추가 및 useWindowStore 타입 참조
