+++ src/components/minihome/guestbook/GuestBook.tsx
@@ -1,5 +1,10 @@
 import { useEffect, useState } from "react";
 
+import type { GuestbookWithComments } from "@/components/minihome/guestbook/types/guestbook.types";
+import Comment from "@/components/minihome/guestbook/ui/Comment";
+import CommentInput from "@/components/minihome/guestbook/ui/CommentInput";
+import Entry from "@/components/minihome/guestbook/ui/Entry";
+import EntryTextArea from "@/components/minihome/guestbook/ui/EntryTextArea";
 import { useAuthStore } from "@/stores/useAuthStore";
 
 import {
@@ -9,11 +14,6 @@ import {
   insertEntry,
   insertReply,
 } from "./api/guestbook";
-import CommentWriteBox from "./CommentWriteBox";
-import GuestBookEntry from "./GuestBookEntry";
-import GuestbookWriteBox from "./GuestbookWriteBox";
-import Reply from "./Reply";
-import type { GuestbookWithComments } from "./types/guestbook.types";
 
 export default function GuestBook({ ownerId }: { ownerId: string | undefined }) {
   const user = useAuthStore((state) => state.user);
@@ -105,7 +105,7 @@ export default function GuestBook({ ownerId }: { ownerId: string | undefined }) 
     );
   };
 
-  const handleEntryWrite = async (content: string) => {
+  const handleEntry = async (content: string) => {
     if (content.trim() === "") return;
     if (!user?.id || !ownerId) return;
 
@@ -123,14 +123,14 @@ export default function GuestBook({ ownerId }: { ownerId: string | undefined }) 
 
   return (
     <div className="flex h-full min-h-0 flex-col p-4">
-      {!isMine && <GuestbookWriteBox onSubmit={handleEntryWrite} />}
+      {!isMine && <EntryTextArea onSubmit={handleEntry} />}
       <div className="flex min-h-0 flex-1 flex-col gap-1">
         <span className="p">전체 방명록 {data.length}</span>
         <div className="bevel-pressed bg-text-invert scrollbar flex min-h-0 flex-1 flex-col gap-2 overflow-y-auto p-3">
           <ul className="flex flex-col gap-2">
             {data.map((entry) => (
               <li key={entry.id}>
-                <GuestBookEntry
+                <Entry
                   id={entry.id}
                   author={entry.author}
                   created_at={entry.created_at}
@@ -139,7 +139,7 @@ export default function GuestBook({ ownerId }: { ownerId: string | undefined }) 
                   onDelete={() => handleEntryDelete(entry.id)}
                 >
                   {entry.comments.length > 0 ? (
-                    <Reply
+                    <Comment
                       id={entry.comments[0].id}
                       commenter={entry.comments[0].commenter}
                       created_at={entry.comments[0].created_at}
@@ -148,9 +148,9 @@ export default function GuestBook({ ownerId }: { ownerId: string | undefined }) 
                       onDelete={() => handleReplyDelete(entry.id, entry.comments[0].id)}
                     />
                   ) : (
-                    isMine && <CommentWriteBox entryId={entry.id} onSubmit={handleReplyWrite} />
+                    isMine && <CommentInput entryId={entry.id} onSubmit={handleReplyWrite} />
                   )}
-                </GuestBookEntry>
+                </Entry>
               </li>
             ))}
           </ul>
