+++ src/components/minihome/guestbook/ui/Comment.tsx
@@ -2,15 +2,20 @@ import dayjs from "dayjs";
 import { X } from "lucide-react";
 
 import Button from "@/components/common/Button/Button";
+import type { GuestbookCommentWithAuthor } from "@/components/minihome/guestbook/types/guestbook.types";
 
-import type { GuestbookCommentWithAuthor } from "./types/guestbook.types";
-
-interface ReplyProps extends GuestbookCommentWithAuthor {
+interface CommentProps extends GuestbookCommentWithAuthor {
   canDelete?: boolean;
   onDelete?: () => void;
 }
 
-export default function Reply({ commenter, created_at, content, canDelete, onDelete }: ReplyProps) {
+export default function Comment({
+  commenter,
+  created_at,
+  content,
+  canDelete,
+  onDelete,
+}: CommentProps) {
   return (
     <div className="border-primary border-l-3 pl-3">
       <div className="flex flex-col gap-2">
@@ -19,8 +24,8 @@ export default function Reply({ commenter, created_at, content, canDelete, onDel
           <div className="flex gap-2">
             <span className="opacity-50">{dayjs(created_at).format("YYYY.MM.DD HH:mm")}</span>
             {canDelete && (
-              <Button onClick={onDelete} composition="iconOnly" size="sm" aria-label="답글 삭제">
-                <X size="1em" strokeWidth={3} />
+              <Button onClick={onDelete} composition="iconOnly" aria-label="답글 삭제">
+                <X className="size-3" strokeWidth={3} />
               </Button>
             )}
           </div>
