+++ src/components/minihome/guestbook/ui/Entry.tsx
@@ -3,23 +3,22 @@ import { X } from "lucide-react";
 import type { ReactNode } from "react";
 
 import Button from "@/components/common/Button/Button";
+import type { GuestBookEntryType } from "@/components/minihome/guestbook/types/guestbook.types";
 
-import type { GuestBookEntryType } from "./types/guestbook.types";
-
-export interface GuestBookEntryProps extends GuestBookEntryType {
+export interface EntryProps extends GuestBookEntryType {
   canDelete?: boolean;
   onDelete?: () => void;
   children?: ReactNode;
 }
 
-export default function GuestBookEntry({
+export default function Entry({
   author,
   created_at,
   content,
   canDelete,
   onDelete,
   children,
-}: GuestBookEntryProps) {
+}: EntryProps) {
   return (
     <div className="bevel-default flex w-full flex-col gap-2 p-4">
       <div className="flex flex-col gap-2">
@@ -28,14 +27,8 @@ export default function GuestBookEntry({
           <div className="flex gap-2">
             <span className="opacity-50">{dayjs(created_at).format("YYYY.MM.DD HH:mm")}</span>
             {canDelete && (
-              <Button
-                onClick={onDelete}
-                composition="iconOnly"
-                size="sm"
-                className="font-bold"
-                aria-label="방명록 삭제"
-              >
-                <X size="1em" strokeWidth={3} />
+              <Button onClick={onDelete} composition="iconOnly" aria-label="방명록 삭제">
+                <X className="size-3" strokeWidth={3} />
               </Button>
             )}
           </div>
