+++ src/config/windowInfo.ts
