+++ src/main.tsx
@@ -3,7 +3,7 @@ import { createRoot } from "react-dom/client";
 import { BrowserRouter } from "react-router";
 
 import App from "@/App";
-import { ThemeProvider } from "@/providers/ThemeProvider";
+import { ThemeProvider } from "@/contexts/ThemeContext";
 import "@/styles/globals.css";
 
 createRoot(document.getElementById("root")!).render(
